settings_microservice/requirements.txt
Flask==1.1.2
Flask-Migrate
connexion[swagger-ui]
sqlalchemy
PyJWT
marshmallow-sqlalchemy
netifaces
python-consul
